feat(ts): components folder
